requirements.txt
opencv-python>=4.8.0
mediapipe>=0.10.0
tensorflow>=2.13.0
numpy>=1.24.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Pillow>=10.0.0
xmltodict>=0.13.0
tqdm>=4.65.0
seaborn>=0.12.0
flask>=2.3.0
flask-socketio>=5.3.0
flask-cors>=4.0.0 